% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ss_range}
\alias{ss_range}
\title{Computes range of sample sizes appropriate for given trial based on
heuristic of diminishing returns}
\usage{
ss_range(
  alpha,
  power,
  maybe_lower_spending_fcn,
  maybe_upper_spending_fcn,
  maybe_look_fractions,
  prop_treated,
  lambda_event_trt,
  lambda_event_ctrl,
  maybe_lambda_dropout,
  enrollment_rates,
  enrollment_times,
  maybe_custom_alpha_spend,
  tol,
  delta,
  min_perc_change
)
}
\description{
Computes range of sample sizes appropriate for given trial based on
heuristic of diminishing returns
}
