% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ctcompute}
\alias{ctcompute}
\title{Computes characteristics of a clinical trial with specified
power, sample size, etc.}
\usage{
ctcompute(
  n_patients,
  alpha,
  power,
  maybe_lower_spending_fcn,
  maybe_upper_spending_fcn,
  maybe_look_fractions,
  prop_treated,
  lambda_event_trt,
  lambda_event_ctrl,
  maybe_lambda_dropout,
  enrollment_rates,
  enrollment_times,
  maybe_custom_alpha_spend,
  r,
  tol
)
}
\description{
Computes characteristics of a clinical trial with specified
power, sample size, etc.
}
